+++ src/App.jsx
@@ -7,8 +7,7 @@ import Guide from "./pages/guide/Guide";
 import { RouterProvider } from "react-router-dom";
 import Layout from "./shared/Layout";
 import { useEffect, useState } from "react";
-import { auth } from "./firebase";
-import styled from "styled-components";
+import { auth } from "./shared/firebase";
 import LoadingScreen from "./shared/components/LoadingScreen";
 import ProtectedRoute from "./shared/components/ProtectedRoute";
 import GlobalStyle from "./shared/components/styles/GlobalStyle";
