+++ src/pages/salaryAdjustment/SalaryAdjustment.jsx
@@ -1,4 +1,4 @@
-import PageTitle from "../../shared/components/titles/PageTitle";
+import PageTitle from "../../shared/components/PageTitle";
 
 const SalaryAdjustment = () => {
   return <PageTitle title="정정 신청 / 내역" subtitle="정정 내역" />;
