+++ src/shared/components/ProtectedRoute.jsx
@@ -1,5 +1,5 @@
 import { Navigate } from "react-router-dom";
-import { auth } from "../../firebase";
+import { auth } from "../firebase";
 
 const ProtectedRoute = ({ children }) => {
   const user = auth.currentUser;
