+++ src/shared/components/modal/Modal.jsx
@@ -1,6 +1,6 @@
 import styled from "styled-components";
-import Button from "../../components/button/Button";
-import CancelBtn from "../../../assets/images/cancel.svg";
+import Button from "../Button";
+import CancelBtn from "/images/cancel.svg";
 
 const Overlay = styled.div`
   position: fixed;
